refactor(app): add OverridePathToPython component and its test to reduce AdvanceSettings code
add OverridePathToPython component and its test to reduce AdvanceSettings code

close RAUT-938 partially
